// src/server/api/models/banking/models/AccountGroup.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { AccountGroupStatus } from './AccountGroupStatus';
import type { AccountGroupType } from './AccountGroupType';
import type { LocalDateTime } from './LocalDateTime';
import type { UUID } from './UUID';

export type AccountGroup = {
    /**
     * The unique ID for the record. This is the ID that is used in all services and related records when referencing the record.
     */
    id: UUID;
    /**
     * The ID of the user that created the record.
     */
    createdBy: string;
    /**
     * When the record was created.
     */
    createdAt: LocalDateTime;
    /**
     * The ID of the user that last modified the record.
     */
    updatedBy: string;
    /**
     * When the record was last modified.
     */
    updatedAt: LocalDateTime;
    /**
     * Indicates if this is a deleted record
     */
    deleted: boolean;
    /**
     * A unique code generated by the system
     */
    code: string;
    /**
     * The group description
     */
    label: string;
    /**
     * The description of the group
     */
    notes?: string | null;
    /**
     * The status of the group
     */
    status: AccountGroupStatus;
    /**
     * The type of the group
     */
    type: AccountGroupType;
};